move useFilters to its own file
